requirements.txt
requests==2.21.0
arxiv==1.4.0
python-dotenv==0.15.0
